Beginning of a game
The idea is for the player to implement the brain of a jeep that goes
into the desert.

The goal is located 1000 kms into the desert. The jeep can take 200
liters of gasoline at a time, and can travel 2.5 kilometers per liter
of gasoline. This will only take the jeep half of the way, so the
player is allowed to store any amount of gasoline currently in the
tank at any point of the way, then go back to the starting point or to
a point with previously stored gasoline to refuel.

The goal is to get to the goal using as little fuel as possible.

 START            1/4          half way           3/4             GOAL
   |---------------|---------------|---------------|---------------|
infinite                      200 liters                       1000 km
gasoline                   here and you win